Massive overhaul phase 1
- updated API library to avoid forthcoming `User.ID` `int32` overflow
- no more config.yml, see README.md on the idea how to migrate
- got rid of dead code
- dockerized
- fixed bug of incomplete cleanup in common groups
- fixed main thread race condition
- a LOT of other little fixes and refactoring

+++ internal/db/dependencies.go
@@ -1,11 +1,11 @@
 package db
 
 type Client interface {
-	GetUserMeta(userID int) (*UserMeta, error)
+	GetUserMeta(userID int64) (*UserMeta, error)
 	UpsertUserMeta(chat *UserMeta) error
 	GetChatMeta(chatID int64) (*ChatMeta, error)
 	UpsertChatMeta(chat *ChatMeta) error
-	GetCharadeScore(chatID int64, userID int) (*CharadeScore, error)
+	GetCharadeScore(chatID int64, userID int64) (*CharadeScore, error)
 	GetCharadeStats(chatID int64) ([]*CharadeScore, error)
-	AddCharadeScore(chatID int64, userID int) (*CharadeScore, error)
+	AddCharadeScore(chatID int64, userID int64) (*CharadeScore, error)
 }
